backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.3.0
pandas
numpy<2.0.0
pytz==2023.3
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
py_clob_client
eth_account
web3
